requirements.txt
setuptools~=51.1.2
PyYAML~=5.3.1
pandas~=1.1.5
apache-airflow~=2.1.1
pendulum~=2.1.2
pytest~=6.2.4

pandera~=0.6.5
